+++ CHANGELOG.md
@@ -5,6 +5,7 @@ UNRELEASED
 * [ [#1883](https://github.com/digitalfabrik/integreat-cms/issues/1883) ] Introduce new status for automatic translations
 * [ [#1993](https://github.com/digitalfabrik/integreat-cms/issues/1993) ] Skip duplicate page translations in XLIFF file
 * [ [#1769](https://github.com/digitalfabrik/integreat-cms/issues/1769) ] Evenly distribute page, event and poi form sidebar boxes
+* [ [#2033](https://github.com/digitalfabrik/integreat-cms/issues/2033) ] Add user setting to optionally enable the automatic distribution of sidebar boxes
 * [ [#1876](https://github.com/digitalfabrik/integreat-cms/issues/1876) ] Make items in the sidebar of poi and event form toggleable
 * [ [#2023](https://github.com/digitalfabrik/integreat-cms/issues/2023) ] Rephrase texts in text understandability-box
 * [ [#1552](https://github.com/digitalfabrik/integreat-cms/issues/1552) ] Make slugfield scrollable